+++ src/scripts/main.js
@@ -18,3 +18,12 @@ document.addEventListener('DOMContentLoaded', () => {
   }
   initFormInteractions(document.querySelector('form'), table);
 });
+
+export function validateOffice(office) {
+  if (!office) {
+    return {
+      title: 'Office',
+      message: 'Office field cannot be empty.',
+    };
+  }
+}
