--- src/scripts/validators/validateAge.js
@@ -1,8 +1,0 @@
-export function validateAge(age) {
-  if (!age || age < 18 || age > 90 || !Number(age)) {
-    return {
-      title: 'Age',
-      message: 'Age must be between 18 and 90 years old.',
-    };
-  }
-}
