--- src/scripts/validators/validateOffice.js
@@ -1,8 +1,0 @@
-export function validateOffice(office) {
-  if (!office) {
-    return {
-      title: 'Office',
-      message: 'Office field cannot be empty.',
-    };
-  }
-}
